backend/requirements.txt
fastapi
uvicorn[standard]
SPARQLWrapper
psycopg2-binary